+++ cloud-spanner-r2dbc/src/main/java/com/google/cloud/spanner/r2dbc/SpannerStatement.java
@@ -161,4 +161,8 @@ public class SpannerStatement implements Statement {
     }
   }
 
+  @Override
+  public Statement fetchSize(int rows) {
+    return null;
+  }
 }
